Rename example modules filenames
