+++ examples/modules/__init__.py
@@ -1,6 +1,6 @@
-from .a import ModuleA
-from .b import ModuleB
-from .c import ModuleC
-from .d import ModuleD
+from .module_a import ModuleA
+from .module_b import ModuleB
+from .module_c import ModuleC
+from .module_d import ModuleD
 
 __all__ = ['ModuleA', 'ModuleB', 'ModuleC', 'ModuleD']
