+++ examples/modules/module_c.py
